Move sqlite specific add splitter methods to sqlite datasource
